% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{ReproducibilityDec2022-package}
\alias{ReproducibilityDec2022}
\alias{ReproducibilityDec2022-package}
\title{ReproducibilityDec2022: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
